requirements.txt
aiofiles
ansible
ansible-core
beautifulsoup4
bs4
certifi
cffi
charset-normalizer
cryptography
idna
Jinja2
MarkupSafe
packaging
pexpect
ptyprocess
pycparser
PyYAML
requests
resolvelib
ruamel.yaml
ruamel.yaml.clib
soupsieve
tqdm
urllib3
aiohttp
